requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
boto3>=1.34.0
python-dotenv==1.0.0
pydantic==2.5.3
sqlalchemy==2.0.25
email-validator>=2.0.0
psycopg2-binary>=2.9.9
